requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==11.1.0
requests==2.32.3
setuptools==75.8.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
